requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
torchinfo>=1.8.0
onnx>=1.16.0
onnxruntime-gpu>=1.18.0
onnx2tf>=1.20.0
tensorflow-cpu>=2.15.0
sng4onnx>=1.0.1
